cpu-tegra.c: default both force_policy_max and edp_enable to "1"
force_policy_max is available in sysfs under the cpu-tegra module if a
ROM dev would want to override this for some reason. This was causing
the cpu to spike above the frequency limit set from userland or
during ramdisk init.

edp_enable (a static variable not in sysfs) will be flipped
immediately to "0" unless the cpu is very warm on boot.